move mafiaBot to its file
